Support upfront 12x limits for yearly plans

+++ apps/web/app/(ee)/api/cron/usage/utils.ts
@@ -11,6 +11,7 @@ import {
   getAdjustedBillingCycleStart,
   log,
 } from "@dub/utils";
+import { getMonth, getYear } from "date-fns";
 
 const limit = 100;
 
@@ -62,50 +63,61 @@ export const updateUsage = async () => {
   // adjustedBillingCycleStart that matches today's date
   const billingReset = workspaces.filter(
     ({ billingCycleStart }) =>
-      getAdjustedBillingCycleStart(billingCycleStart as number) ===
-      new Date().getDate(),
+      getAdjustedBillingCycleStart(billingCycleStart) === new Date().getDate(),
   );
 
   // Reset usage and alert emails for the billingReset workspaces
   // also send 30-day summary email
   await Promise.allSettled(
     billingReset.map(async (workspace) => {
-      const { plan, usage, usageLimit } = workspace;
+      const { usage, usageLimit, plan, planPeriod, billingCycleEndsAt } =
+        workspace;
 
       /* 
         We only reset clicks usage if it's not over usageLimit by:
         - 4x for free plan (4K clicks)
         - 2x for all other plans
       */
-
-      const resetUsage =
+      let resetUsage =
         plan === "free" ? usage <= usageLimit * 4 : usage <= usageLimit * 2;
 
-      await prisma.project.update({
-        where: {
-          id: workspace.id,
-        },
-        data: {
-          ...(resetUsage && {
+      // if yearly plan, we skip resetting usage if the billing cycle end date is not this year+month
+      if (
+        planPeriod === "yearly" &&
+        billingCycleEndsAt &&
+        !(
+          getYear(billingCycleEndsAt) === getYear(new Date()) &&
+          getMonth(billingCycleEndsAt) === getMonth(new Date())
+        )
+      ) {
+        resetUsage = false;
+      }
+
+      if (resetUsage) {
+        await prisma.project.update({
+          where: {
+            id: workspace.id,
+          },
+          data: {
             usage: 0,
-          }),
-          linksUsage: 0,
-          payoutsUsage: 0,
-          aiUsage: 0,
-          sentEmails: {
-            deleteMany: {
-              type: {
-                in: [
-                  "firstUsageLimitEmail",
-                  "secondUsageLimitEmail",
-                  "firstLinksLimitEmail",
-                  "secondLinksLimitEmail",
-                ],
+            linksUsage: 0,
+            payoutsUsage: 0,
+            aiUsage: 0,
+            sentEmails: {
+              deleteMany: {
+                type: {
+                  in: [
+                    "firstUsageLimitEmail",
+                    "secondUsageLimitEmail",
+                    "firstLinksLimitEmail",
+                    "secondLinksLimitEmail",
+                  ],
+                },
               },
             },
           },
-        },
-      });
+        });
+      }
 
       /* Only send the 30-day summary email if:
          - the workspace has at least 1 link click
